+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
-- Use paket, FAKE and update to dotnet core 3.0
 - Initial implementation
+- Use paket, FAKE and update to dotnet core 3.0
+- Automatically exclude a target directory
+- Add `force` option